10 levels, boss enemy, game states

+++ src/FastEnemy.java
@@ -1,14 +1,14 @@
 import java.awt.*;
 public class FastEnemy extends GameObject{
     private Handler handler;
-    public FastEnemy(int x, int y, ID id, Handler handler) {
+    public FastEnemy(float x, float y, ID id, Handler handler) {
         super(x, y, id);
         velX = 8;
         velY = 8;
         this.handler = handler;
     }
     public Rectangle getBounds(){
-        return new Rectangle(x,y,16,16);
+        return new Rectangle((int) x,(int) y,16,16);
     }
     public void tick() {
         x += velX;
@@ -17,11 +17,11 @@ public class FastEnemy extends GameObject{
         if(y <= 0 || y >= Game.HEIGHT - 32) velY *= -1;
         if(x <= 0 || x >= Game.WIDTH - 32) velX *= -1;
 
-        handler.addObject(new Trail(x, y , ID.Trail, Color.cyan, 16, 16,0.05f, handler));
+        handler.addObject(new Trail((int) x, (int) y , ID.Trail, Color.cyan, 16, 16,0.05f, handler));
     }
 
     public void render(Graphics g) {
         g.setColor(Color.cyan);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int) x,(int) y,16,16);
     }
 }
