+++ src/GameObject.java
@@ -1,11 +1,11 @@
 import java.awt.*;
 
 public abstract class GameObject {
-    protected int x,y;
+    protected float x,y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
-    public GameObject(int x, int y, ID id){
+    public GameObject(float x, float y, ID id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -20,10 +20,10 @@ public abstract class GameObject {
     public void setY(int y){
         this.y = y;
     }
-    public int getX(){
+    public float getX(){
         return x;
     }
-    public int getY(){
+    public float getY(){
         return y;
     }
     public void setId(ID id){
@@ -32,16 +32,16 @@ public abstract class GameObject {
     public ID getId(){
         return id;
     }
-    public void setVelX(int velX){
+    public void setVelX(float velX){
         this.velX = velX;
     }
-    public void setVelY(int velY){
+    public void setVelY(float velY){
         this.velY = velY;
     }
-    public int getVelX(){
+    public float getVelX(){
         return velX;
     }
-    public int getVelY(){
+    public float getVelY(){
         return velY;
     }
 }
