+++ src/HUD.java
@@ -9,8 +9,8 @@ public class HUD {
 
     public void tick(){
         // HEALTH --;
-        HEALTH = Game.clamp(HEALTH, 0, 100);
-        greenValue = Game.clamp(greenValue, 0, 255);
+        HEALTH = (int) Game.clamp(HEALTH, 0, 100);
+        greenValue = (int) Game.clamp(greenValue, 0, 255);
         greenValue = HEALTH * 2;
         score++;
     }
